refactor: MoneyProvider Layout.jsx에서 분리
