+++ src/pages/Wallet/Wallet.jsx
@@ -2,20 +2,15 @@ import { useContext } from "react";
 
 import MoneyItem from "components/Wallet/MoneyItem/MoneyItem";
 import TotalMoneyArea from "components/Wallet/TotalMoneyArea/TotalMoneyArea";
-import { MoneyContext } from "pages/Layout/Layout";
+import { MoneyContext } from "contexts/moneyContext";
 
 import Wrapper from "./Wallet.styled";
 
 const Wallet = () => {
   const { cashData } = useContext(MoneyContext);
 
-  // const walletItems = cashData.map(({ money, count }) => {
-  //   return <MoneyItem key={money} money={money} count={count} />;
-  // });
-
   return (
     <Wrapper>
-      {/* <ul>{walletItems}</ul> */}
       <ul>
         {cashData.map(({ money, count }) => (
           <MoneyItem key={money} money={money} count={count} />
